beacon/light: moved canonicalStore to separate file
